Add files via upload
For reviewing a place a user has to Sign-up if he/she is new to the website or simply can log in if an existing user. After that they will be welcomed with a world map where a user can review any place and also add images by simply double-clicking on the place which they want to review. After double-clicking a form will open where the user has to enter the description of the place and give it a rating and adding an image is optional. Once submitted the review will be visible to everyone who visits the site. One can also find a place using the search option at the top. 

TECHNOLOGIES USED:-
1)  ReactJS(FrontEnd)
2)  NodeJS(BackEnd)
3)  MongoDB(Storage)